+++ common/src/main/java/com/st0x0ef/stellaris/common/data/recipes/WaterSeparatorRecipe.java
@@ -22,7 +22,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
-public record WaterSeparatorRecipe(FluidStack ingredientStack, List<FluidStack> resultStacks, boolean isMb, long energy) implements Recipe<FluidInput> {
+public record WaterSeparatorRecipe(FluidStack ingredientStack, List<FluidStack> resultStacks, boolean isMb, int energy) implements Recipe<FluidInput> {
 
     @Override
     public boolean matches(FluidInput container, Level level) {
@@ -61,7 +61,7 @@ public record WaterSeparatorRecipe(FluidStack ingredientStack, List<FluidStack> 
                 FluidStack.CODEC.fieldOf("ingredient").forGetter(WaterSeparatorRecipe::ingredientStack),
                 FluidStack.CODEC.listOf(1, 2).fieldOf("results").forGetter(WaterSeparatorRecipe::resultStacks),
                 Codec.BOOL.optionalFieldOf("isFluidMB").forGetter(recipe -> Optional.of(recipe.isMb)),
-                Codec.LONG.fieldOf("energy").forGetter(WaterSeparatorRecipe::energy)
+                Codec.INT.fieldOf("energy").forGetter(WaterSeparatorRecipe::energy)
         ).apply(instance, (ingredientStack, resultStacks, isFluidMb, energy) -> {
             boolean isMb = isFluidMb.orElse(true);
             convertFluidStack(ingredientStack, isMb);
@@ -76,7 +76,7 @@ public record WaterSeparatorRecipe(FluidStack ingredientStack, List<FluidStack> 
             FLUID_STACK_LIST_STREAM_CODEC.encode(buf, recipe.resultStacks);
             buf.writeBoolean(recipe.isMb);
             buf.writeLong(recipe.energy);
-        }, buf -> new WaterSeparatorRecipe(FluidStack.read(buf), FLUID_STACK_LIST_STREAM_CODEC.decode(buf), buf.readBoolean(), buf.readLong()));
+        }, buf -> new WaterSeparatorRecipe(FluidStack.read(buf), FLUID_STACK_LIST_STREAM_CODEC.decode(buf), buf.readBoolean(), buf.readInt()));
 
         public static void convertFluidStack(FluidStack stack, boolean isMb) {
             if (isMb) {
